- merge db/nodes.ts into utils/nodes.ts
